Custom pupil mask + input_frame_count + roi_via_matrix
CAW has been updated so that it can operate with any custom pupil mask (mask must be the same for all SHWFSs). There is also now the option to input a frame count for the SHWFS iterations (useful if frames are missed, jumbled etc). The code can also be directed to calculate the ROI via the covariance matrix. This is the fastest technique for AO systems with large SHWFS dimensions e.g. AOF.